requirements.txt


python-dotenv